% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapBooklet.R
\docType{package}
\name{MapBooklet}
\alias{MapBooklet}
\alias{MapBooklet-package}
\title{R package for generating map booklets}
\description{
This R package was designed for a automized generation of map booklets based on user ggplots.
}
\details{
Functions included in this package are:
\itemize{
  \item \code{\link[MapBooklet]{FishnetFunction}}
  \item \code{\link[MapBooklet]{BookletMaker}}
}
}
\author{
Marius Philipp \email{marius.philipp@stud-mail.uni-wuerzburg.de}
}
